work

+++ desktop_version/build_psp.sh
@@ -2,33 +2,40 @@
 
 # PSPDEV=$(psp-config -d)
 
-rm -rf build/*
-rm Makefile
+BUILD=true
+LINK=true
 
-mkdir -p build
+if [ "$BUILD" = true ]; then
+    rm -rf build/*
+    rm Makefile
 
-cd build
+    mkdir -p build
 
-# -DCMAKE_VERBOSE_MAKEFILE=1 \
-cmake .. \
-    -DSDL_INCLUDE_DIRS="$PSPDEV/psp/include/SDL" \
-    -DSDL_LIBRARIES="$PSPDEV/psp/lib/libSDL.a;$PSPDEV/psp/lib/libSDLmain.a;$PSPDEV/psp/lib/libSDL_mixer.a" \
-    || exit 1
+    cd build
 
-cd ../
+    # -DCMAKE_VERBOSE_MAKEFILE=1 \
+    cmake .. \
+        -DSDL_INCLUDE_DIRS="$PSPDEV/psp/include/SDL" \
+        -DSDL_LIBRARIES="$PSPDEV/psp/lib/libSDL.a;$PSPDEV/psp/lib/libSDLmain.a;$PSPDEV/psp/lib/libSDL_mixer.a" \
+        || exit 1
 
-cmake --build build
+    cd ../
+
+    cmake --build build
+fi
 
 
-echo
-echo "-------REAL LINKER HERE-------"
-echo
+if [ "$LINK" = true ]; then
+    echo
+    echo "-------REAL LINKER HERE-------"
+    echo
 
-cd build
+    cd build
 
-set -ex
+    set -ex
 
-psp-gcc -g -G0 -D_GNU_SOURCE=1 -D__POSIX_C_VISIBLE=200809 -L$PSPDEV/psp/sdk/lib -L$PSPDEV/psp/lib -o VVVVVV ./CMakeFiles/VVVVVV.dir/src/*.obj -L./ -llodepng-static -ltinyxml-static -lphysfs-static -lstdc++ -lSDL_mixer -lSDL -lGL -lmikmod -lpspvfpu -lpspvram -lvorbisidec -lpthread-psp -lc  -lpsphprm -lpspdebug -lpspgu -lpspctrl -lpspdisplay -lpspge -lpspsdk -lpsprtc -lpspaudio -lpspnet -lpspnet_inet -lpsppower -lpsputility -lpspuser -lpspkernel -lm -specs=$PSPDEV/psp/sdk/lib/prxspecs -Wl,-T$PSPDEV/psp/sdk/lib/linkfile.prx,-q
+    psp-gcc -g -G0 -D_GNU_SOURCE=1 -D__POSIX_C_VISIBLE=200809 -L$PSPDEV/psp/sdk/lib -L$PSPDEV/psp/lib -o VVVVVV ./CMakeFiles/VVVVVV.dir/src/*.obj -L./ -llodepng-static -ltinyxml-static -lphysfs-static -lstdc++ -lSDL_mixer -lSDL -lGL -lmikmod -lpspvfpu -lpspvram -lvorbisidec -lpthread-psp -lc  -lpsphprm -lpspdebug -lpspgu -lpspctrl -lpspdisplay -lpspge -lpspsdk -lpsprtc -lpspaudio -lpspnet -lpspnet_inet -lpsppower -lpsputility -lpspuser -lpspkernel -lm -specs=$PSPDEV/psp/sdk/lib/prxspecs -Wl,-T$PSPDEV/psp/sdk/lib/linkfile.prx,-q
 
-psp-fixup-imports VVVVVV
-psp-prxgen VVVVVV VVVVVV.prx
+    psp-fixup-imports VVVVVV
+    psp-prxgen VVVVVV VVVVVV.prx
+fi
